pyproject.toml
[tool.poetry]
name = "bloom-cli"
version = "0.0.1"
description = "A simple CLI to make experimenting with BLOOM easier."
authors = ["Lawrence Stewart"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "3.8.10"
certifi = "2022.6.15"
charset-normalizer = "2.1.0"
click = "8.1.3"
colorama = "0.4.5"
commonmark = "0.9.1"
idna = "3.3"
Pygments = "2.13.0"
python-dotenv = "0.20.0"
requests = "2.28.1"
rich = "12.5.1"
shellingham = "1.5.0"
typer = {extras = ["all"], version = "0.6.1"}
typing-extensions = "4.3.0"
urllib3 = "1.26.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
bloom-cli = "bloom_cli.main:app"

